refactor: simplify and deduplicate post-migration code
- is_404 + is_409 merged into is_http_status(err, code): same pattern,
  one function
- FakeDockerClient.remove_{container,volume,network}: three identical
  or_else "no such" blocks extracted to ignore_if_missing helper
- gc_orphaned_resources: four sequential remove awaits replaced with
  tokio::join! so the four independent Docker calls run concurrently
- NoOpDocker + FakeRunner extracted from amp_launch.rs and codex_launch.rs
  into tests/common/mod.rs; both test files reduced by ~120 lines
