+++ tests/codex_launch.rs
@@ -1,136 +1,14 @@
+mod common;
+
+use common::{FakeRunner, NoOpDocker};
 use jackin::agent::Agent;
 use jackin::config::AppConfig;
-use jackin::docker::{CommandRunner, RunOptions};
-use jackin::docker_client::{
-    ContainerRow, ContainerSpec, ContainerState, DockerApi, NetworkRow, RemoveImageOutcome,
-};
 use jackin::isolation::MountIsolation;
 use jackin::paths::JackinPaths;
 use jackin::runtime::{LoadOptions, load_role};
 use jackin::selector::RoleSelector;
 use jackin::workspace::{MountConfig, ResolvedWorkspace};
-use std::collections::{HashMap, VecDeque};
-use std::path::Path;
 use tempfile::tempdir;
-
-// Minimal DockerApi stub for integration tests: all GC/inspect calls return
-// empty results so load_role proceeds as if no containers exist.
-struct NoOpDocker;
-
-impl DockerApi for NoOpDocker {
-    async fn inspect_container_state(&self, _name: &str) -> ContainerState {
-        ContainerState::NotFound
-    }
-    async fn remove_container(&self, _name: &str) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn list_containers(
-        &self,
-        _label_filters: &[&str],
-        _all: bool,
-    ) -> anyhow::Result<Vec<ContainerRow>> {
-        Ok(vec![])
-    }
-    async fn create_container(&self, _name: &str, _spec: ContainerSpec) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn start_container(&self, _name: &str) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn remove_volume(&self, _name: &str) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn create_network(
-        &self,
-        _name: &str,
-        _labels: HashMap<String, String>,
-    ) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn remove_network(&self, _name: &str) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn list_networks(&self, _label_filters: &[&str]) -> anyhow::Result<Vec<NetworkRow>> {
-        Ok(vec![])
-    }
-    async fn list_image_tags(&self, _reference_filter: &str) -> anyhow::Result<Vec<String>> {
-        Ok(vec![])
-    }
-    async fn remove_image(&self, _name: &str) -> anyhow::Result<RemoveImageOutcome> {
-        Ok(RemoveImageOutcome::NotFound)
-    }
-    async fn inspect_image_label(
-        &self,
-        _image: &str,
-        _label: &str,
-    ) -> anyhow::Result<Option<String>> {
-        Ok(None)
-    }
-    async fn pull_image(&self, _image: &str, _debug: bool) -> anyhow::Result<()> {
-        Ok(())
-    }
-    async fn exec_capture(&self, _container: &str, _cmd: &[&str]) -> anyhow::Result<String> {
-        Ok(String::new())
-    }
-}
-
-#[derive(Default)]
-struct FakeRunner {
-    recorded: Vec<String>,
-    capture_queue: VecDeque<String>,
-}
-
-impl FakeRunner {
-    fn for_load_agent(outputs: impl IntoIterator<Item = String>) -> Self {
-        // Preamble: 4 identity lookups (git config user.name, user.email, id -u, id -g).
-        // GC now uses DockerApi, not CommandRunner, so it no longer counts.
-        let mut capture_queue = VecDeque::new();
-        for _ in 0..4 {
-            capture_queue.push_back(String::new());
-        }
-        capture_queue.extend(outputs);
-        Self {
-            recorded: Vec::new(),
-            capture_queue,
-        }
-    }
-}
-
-impl CommandRunner for FakeRunner {
-    async fn run(
-        &mut self,
-        program: &str,
-        args: &[&str],
-        _cwd: Option<&Path>,
-        _opts: &RunOptions,
-    ) -> anyhow::Result<()> {
-        self.recorded.push(format!("{program} {}", args.join(" ")));
-        Ok(())
-    }
-
-    async fn capture(
-        &mut self,
-        program: &str,
-        args: &[&str],
-        _cwd: Option<&Path>,
-    ) -> anyhow::Result<String> {
-        self.recorded.push(format!("{program} {}", args.join(" ")));
-        Ok(self.capture_queue.pop_front().unwrap_or_default())
-    }
-
-    async fn capture_secret(
-        &mut self,
-        program: &str,
-        args: &[&str],
-        cwd: Option<&Path>,
-    ) -> anyhow::Result<String> {
-        // Delegates to `capture` and consumes one queue slot. Always provision
-        // one entry per expected `capture_secret` call (e.g. `gh auth token`
-        // in `resolve_github_token`) and document it above the `for_load_agent`
-        // call — same discipline as for `capture` calls.
-        self.capture(program, args, cwd).await
-    }
-}
 
 #[tokio::test]
 async fn codex_launch_invokes_docker_run_with_codex_agent() {
